Optimise addition of new elements in explicit multisets.
When having explicit multisets, adding a new element is often very slow,
mostly because of overhead in bsearch() and qsort(). This cost may be
significant and this optimisation brings speeds up to around 50-100
times.

This can be very significant when doing group by on some column and
having a long tail of groups with low cardinality. E.g. when doing
something similiar to:

   SELECT mygroup, hll_add_agg(hll_hash_any(mykey))
   FROM mytable GROUP BY mygroup ORDER BY mygroup;