Stop generating migration files for each version
PostgreSQL 10 introduced an improvement on extension migration paths. We
no longer need to generate a migration file for each version. It is
sufficient to have an initial script and the migration scripts that are
used to upgrade to the next version.

This commit also fixes an issue where we did not install necessary
migration files on a clean build.